feat: ask which files to symlink

+++ setup.sh
@@ -2,12 +2,20 @@
 
 # RUN THIS INSIDE .git/hooks OF YOUR GIT REPO
 
-cat extra_words >> $HOME/.hunspell_en_US
+cat $HOME/developer/git-hooks/extra_words >> $HOME/.hunspell_en_US
 
-echo "Symlink files? [y/n]"
+echo "Symlink files? [y/N]"
 read run_symlink
 
 if [[ "$run_symlink" == "y" ]]; then
-    ln -s $HOME/developer/git-hooks/commit-msg .git/hooks
-    ln -s $HOME/developer/git-hooks/pre-commit .git/hooks
+    echo "Symlink commit-msg? [y/N]"
+    read prompt
+    if [[ "$prompt" == "y" ]]; then
+        ln -s $HOME/developer/git-hooks/commit-msg .git/hooks
+    fi
+    echo "Symlink pre-commit? [y/N]"
+    read prompt
+    if [[ "$prompt" == "y" ]]; then
+        ln -s $HOME/developer/git-hooks/pre-commit .git/hooks
+    fi
 fi
